Suggested edit from issues/136
It seems to me that when the emulator gives the ".desktop" file a new GUID in its filename every time it runs, it kind of makes things a hassle for people who manually organize or manage their desktop shortcuts on Linux. See https://github.com/NovaSquirrel/Mesen-X/issues/136 for an example of this. (But there's probably someone who knows more about this than I do who can confirm or deny what I'm saying.)

I am not the most familiar with the code for this emulator, but it almost looks like if this proposal is implemented, it could also be helpful to move line 170 of /GUI.NET/Config/PreferenceInfo.cs to after line 72 in GUI.NET/Forms/frmConfigWizard.cs so that it only creates the .desktop file once upon setup and not each time the app is run.